Move StatePolicy to own module
